Enum rusoto_codecommit::PostCommentForPullRequestError

pub enum PostCommentForPullRequestError {
    BeforeCommitIdAndAfterCommitIdAreSame(String),
    ClientRequestTokenRequired(String),
    CommentContentRequired(String),
    CommentContentSizeLimitExceeded(String),
    CommitDoesNotExist(String),
    CommitIdRequired(String),
    EncryptionIntegrityChecksFailed(String),
    EncryptionKeyAccessDenied(String),
    EncryptionKeyDisabled(String),
    EncryptionKeyNotFound(String),
    EncryptionKeyUnavailable(String),
    IdempotencyParameterMismatch(String),
    InvalidClientRequestToken(String),
    InvalidCommitId(String),
    InvalidFileLocation(String),
    InvalidFilePosition(String),
    InvalidPath(String),
    InvalidPullRequestId(String),
    InvalidRelativeFileVersionEnum(String),
    InvalidRepositoryName(String),
    PathDoesNotExist(String),
    PathRequired(String),
    PullRequestDoesNotExist(String),
    PullRequestIdRequired(String),
    RepositoryDoesNotExist(String),
    RepositoryNameRequired(String),
    RepositoryNotAssociatedWithPullRequest(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by PostCommentForPullRequest

Variants
BeforeCommitIdAndAfterCommitIdAreSame(String)

The before commit ID and the after commit ID are the same, which is not valid. The before commit ID and the after commit ID must be different commit IDs.

ClientRequestTokenRequired(String)

A client request token is required. A client request token is an unique, client-generated idempotency token that when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request is received with the same parameters and a token is included, the request will return information about the initial request that used that token.

CommentContentRequired(String)

The comment is empty. You must provide some content for a comment. The content cannot be null.

CommentContentSizeLimitExceeded(String)

The comment is too large. Comments are limited to 1,000 characters.

CommitDoesNotExist(String)

The specified commit does not exist or no commit was specified, and the specified repository has no default branch.

CommitIdRequired(String)

A commit ID was not specified.

EncryptionIntegrityChecksFailed(String)

An encryption integrity check failed.

EncryptionKeyAccessDenied(String)

An encryption key could not be accessed.

EncryptionKeyDisabled(String)

The encryption key is disabled.

EncryptionKeyNotFound(String)

No encryption key was found.

EncryptionKeyUnavailable(String)

The encryption key is not available.

IdempotencyParameterMismatch(String)

The client request token is not valid. Either the token is not in a valid format, or the token has been used in a previous request and cannot be re-used.

InvalidClientRequestToken(String)

The client request token is not valid.

InvalidCommitId(String)

The specified commit ID is not valid.

InvalidFileLocation(String)

The location of the file is not valid. Make sure that you include the extension of the file as well as the file name.

InvalidFilePosition(String)

The position is not valid. Make sure that the line number exists in the version of the file you want to comment on.

InvalidPath(String)

The specified path is not valid.

InvalidPullRequestId(String)

The pull request ID is not valid. Make sure that you have provided the full ID and that the pull request is in the specified repository, and then try again.

InvalidRelativeFileVersionEnum(String)

Either the enum is not in a valid format, or the specified file version enum is not valid in respect to the current file version.

InvalidRepositoryName(String)

At least one specified repository name is not valid.

This exception only occurs when a specified repository name is not valid. Other exceptions occur when a required repository parameter is missing, or when a specified repository does not exist.

PathDoesNotExist(String)

The specified path does not exist.

PathRequired(String)

The filePath for a location cannot be empty or null.

PullRequestDoesNotExist(String)

The pull request ID could not be found. Make sure that you have specified the correct repository name and pull request ID, and then try again.

PullRequestIdRequired(String)

A pull request ID is required, but none was provided.

RepositoryDoesNotExist(String)

The specified repository does not exist.

RepositoryNameRequired(String)

A repository name is required but was not specified.

RepositoryNotAssociatedWithPullRequest(String)

The repository does not contain any pull requests with that pull request ID. Check to make sure you have provided the correct repository name for the pull request.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl PostCommentForPullRequestError

pub fn from_body(body: &str) -> PostCommentForPullRequestError