Enum rusoto_codecommit::PutRepositoryTriggersError

pub enum PutRepositoryTriggersError {
    EncryptionIntegrityChecksFailed(String),
    EncryptionKeyAccessDenied(String),
    EncryptionKeyDisabled(String),
    EncryptionKeyNotFound(String),
    EncryptionKeyUnavailable(String),
    InvalidRepositoryName(String),
    InvalidRepositoryTriggerBranchName(String),
    InvalidRepositoryTriggerCustomData(String),
    InvalidRepositoryTriggerDestinationArn(String),
    InvalidRepositoryTriggerEvents(String),
    InvalidRepositoryTriggerName(String),
    InvalidRepositoryTriggerRegion(String),
    MaximumBranchesExceeded(String),
    MaximumRepositoryTriggersExceeded(String),
    RepositoryDoesNotExist(String),
    RepositoryNameRequired(String),
    RepositoryTriggerBranchNameListRequired(String),
    RepositoryTriggerDestinationArnRequired(String),
    RepositoryTriggerEventsListRequired(String),
    RepositoryTriggerNameRequired(String),
    RepositoryTriggersListRequired(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by PutRepositoryTriggers

Variants
EncryptionIntegrityChecksFailed(String)

An encryption integrity check failed.

EncryptionKeyAccessDenied(String)

An encryption key could not be accessed.

EncryptionKeyDisabled(String)

The encryption key is disabled.

EncryptionKeyNotFound(String)

No encryption key was found.

EncryptionKeyUnavailable(String)

The encryption key is not available.

InvalidRepositoryName(String)

At least one specified repository name is not valid.

This exception only occurs when a specified repository name is not valid. Other exceptions occur when a required repository parameter is missing, or when a specified repository does not exist.

InvalidRepositoryTriggerBranchName(String)

One or more branch names specified for the trigger is not valid.

InvalidRepositoryTriggerCustomData(String)

The custom data provided for the trigger is not valid.

InvalidRepositoryTriggerDestinationArn(String)

The Amazon Resource Name (ARN) for the trigger is not valid for the specified destination. The most common reason for this error is that the ARN does not meet the requirements for the service type.

InvalidRepositoryTriggerEvents(String)

One or more events specified for the trigger is not valid. Check to make sure that all events specified match the requirements for allowed events.

InvalidRepositoryTriggerName(String)

The name of the trigger is not valid.

InvalidRepositoryTriggerRegion(String)

The region for the trigger target does not match the region for the repository. Triggers must be created in the same region as the target for the trigger.

MaximumBranchesExceeded(String)

The number of branches for the trigger was exceeded.

MaximumRepositoryTriggersExceeded(String)

The number of triggers allowed for the repository was exceeded.

RepositoryDoesNotExist(String)

The specified repository does not exist.

RepositoryNameRequired(String)

A repository name is required but was not specified.

RepositoryTriggerBranchNameListRequired(String)

At least one branch name is required but was not specified in the trigger configuration.

RepositoryTriggerDestinationArnRequired(String)

A destination ARN for the target service for the trigger is required but was not specified.

RepositoryTriggerEventsListRequired(String)

At least one event for the trigger is required but was not specified.

RepositoryTriggerNameRequired(String)

A name for the trigger is required but was not specified.

RepositoryTriggersListRequired(String)

The list of triggers for the repository is required but was not specified.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl PutRepositoryTriggersError

fn from_body(body: &str) -> PutRepositoryTriggersError