Crate rusoto_codecommit

AWS CodeCommit
This is the AWS CodeCommit API Reference. This reference provides descriptions of the operations and data types for AWS CodeCommit API along with usage examples.

You can use the AWS CodeCommit API to work with the following objects:

Repositories, by calling the following:

BatchGetRepositories, which returns information about one or more repositories associated with your AWS account

CreateRepository, which creates an AWS CodeCommit repository

DeleteRepository, which deletes an AWS CodeCommit repository

GetRepository, which returns information about a specified repository

ListRepositories, which lists all AWS CodeCommit repositories associated with your AWS account

UpdateRepositoryDescription, which sets or updates the description of the repository

UpdateRepositoryName, which changes the name of the repository. If you change the name of a repository, no other users of that repository will be able to access it until you send them the new HTTPS or SSH URL to use.

Branches, by calling the following:

CreateBranch, which creates a new branch in a specified repository

GetBranch, which returns information about a specified branch

ListBranches, which lists all branches for a specified repository

UpdateDefaultBranch, which changes the default branch for a repository

Information about committed code in a repository, by calling the following:

GetBlob, which returns the base-64 encoded content of an individual Git blob object within a repository

GetCommit, which returns information about a commit, including commit messages and author and committer information

GetDifferences, which returns information about the differences in a valid commit specifier (such as a branch, tag, HEAD, commit ID or other fully qualified reference)

Triggers, by calling the following:

GetRepositoryTriggers, which returns information about triggers configured for a repository

PutRepositoryTriggers, which replaces all triggers for a repository and can be used to create or delete triggers

TestRepositoryTriggers, which tests the functionality of a repository trigger by sending data to the trigger target

For information about how to use AWS CodeCommit, see the AWS CodeCommit User Guide.

If you're using the service, you're probably looking for CodeCommitClient and CodeCommit.

Structs

BatchGetRepositoriesInput
Represents the input of a batch get repositories operation.

BatchGetRepositoriesOutput
Represents the output of a batch get repositories operation.

BlobMetadata
Returns information about a specific Git blob object.

BranchInfo
Returns information about a branch.

CodeCommitClient
A client for the CodeCommit API.

Commit
Returns information about a specific commit.

CreateBranchInput
Represents the input of a create branch operation.

CreateRepositoryInput
Represents the input of a create repository operation.

CreateRepositoryOutput
Represents the output of a create repository operation.

DeleteRepositoryInput
Represents the input of a delete repository operation.

DeleteRepositoryOutput
Represents the output of a delete repository operation.

Difference
Returns information about a set of differences for a commit specifier.

GetBlobInput
Represents the input of a get blob operation.

GetBlobOutput
Represents the output of a get blob operation.

GetBranchInput
Represents the input of a get branch operation.

GetBranchOutput
Represents the output of a get branch operation.

GetCommitInput
Represents the input of a get commit operation.

GetCommitOutput
Represents the output of a get commit operation.

GetDifferencesInput

GetDifferencesOutput

GetRepositoryInput
Represents the input of a get repository operation.

GetRepositoryOutput
Represents the output of a get repository operation.

GetRepositoryTriggersInput
Represents the input of a get repository triggers operation.

GetRepositoryTriggersOutput
Represents the output of a get repository triggers operation.

ListBranchesInput
Represents the input of a list branches operation.

ListBranchesOutput
Represents the output of a list branches operation.

ListRepositoriesInput
Represents the input of a list repositories operation.

ListRepositoriesOutput
Represents the output of a list repositories operation.

PutRepositoryTriggersInput
Represents the input ofa put repository triggers operation.

PutRepositoryTriggersOutput
Represents the output of a put repository triggers operation.

RepositoryMetadata
Information about a repository.

RepositoryNameIdPair
Information about a repository name and ID.

RepositoryTrigger
Information about a trigger for a repository.

RepositoryTriggerExecutionFailure
A trigger failed to run.

TestRepositoryTriggersInput
Represents the input of a test repository triggers operation.

TestRepositoryTriggersOutput
Represents the output of a test repository triggers operation.

UpdateDefaultBranchInput
Represents the input of an update default branch operation.

UpdateRepositoryDescriptionInput
Represents the input of an update repository description operation.

UpdateRepositoryNameInput
Represents the input of an update repository description operation.

UserInfo
Information about the user who made a specified commit.

Enums

BatchGetRepositoriesError
Errors returned by BatchGetRepositories

CreateBranchError
Errors returned by CreateBranch

CreateRepositoryError
Errors returned by CreateRepository

DeleteRepositoryError
Errors returned by DeleteRepository

GetBlobError
Errors returned by GetBlob

GetBranchError
Errors returned by GetBranch

GetCommitError
Errors returned by GetCommit

GetDifferencesError
Errors returned by GetDifferences

GetRepositoryError
Errors returned by GetRepository

GetRepositoryTriggersError
Errors returned by GetRepositoryTriggers

ListBranchesError
Errors returned by ListBranches

ListRepositoriesError
Errors returned by ListRepositories

PutRepositoryTriggersError
Errors returned by PutRepositoryTriggers

TestRepositoryTriggersError
Errors returned by TestRepositoryTriggers

UpdateDefaultBranchError
Errors returned by UpdateDefaultBranch

UpdateRepositoryDescriptionError
Errors returned by UpdateRepositoryDescription

UpdateRepositoryNameError
Errors returned by UpdateRepositoryName

Traits

CodeCommit
Trait representing the capabilities of the CodeCommit API. CodeCommit clients implement this trait.