Crate rusoto_codecommit

AWS CodeCommit
This is the AWS CodeCommit API Reference. This reference provides descriptions of the operations and data types for AWS CodeCommit API along with usage examples.

You can use the AWS CodeCommit API to work with the following objects:

Repositories, by calling the following:

BatchGetRepositories, which returns information about one or more repositories associated with your AWS account.

CreateRepository, which creates an AWS CodeCommit repository.

DeleteRepository, which deletes an AWS CodeCommit repository.

GetRepository, which returns information about a specified repository.

ListRepositories, which lists all AWS CodeCommit repositories associated with your AWS account.

UpdateRepositoryDescription, which sets or updates the description of the repository.

UpdateRepositoryName, which changes the name of the repository. If you change the name of a repository, no other users of that repository will be able to access it until you send them the new HTTPS or SSH URL to use.

Branches, by calling the following:

CreateBranch, which creates a new branch in a specified repository.

DeleteBranch, which deletes the specified branch in a repository unless it is the default branch.

GetBranch, which returns information about a specified branch.

ListBranches, which lists all branches for a specified repository.

UpdateDefaultBranch, which changes the default branch for a repository.

Files, by calling the following:

PutFile, which adds or modifies a file in a specified repository and branch.

Information about committed code in a repository, by calling the following:

GetBlob, which returns the base-64 encoded content of an individual Git blob object within a repository.

GetCommit, which returns information about a commit, including commit messages and author and committer information.

GetDifferences, which returns information about the differences in a valid commit specifier (such as a branch, tag, HEAD, commit ID or other fully qualified reference).

Pull requests, by calling the following:

CreatePullRequest, which creates a pull request in a specified repository.

DescribePullRequestEvents, which returns information about one or more pull request events.

GetCommentsForPullRequest, which returns information about comments on a specified pull request.

GetMergeConflicts, which returns information about merge conflicts between the source and destination branch in a pull request.

GetPullRequest, which returns information about a specified pull request.

ListPullRequests, which lists all pull requests for a repository.

MergePullRequestByFastForward, which merges the source destination branch of a pull request into the specified destination branch for that pull request using the fast-forward merge option.

PostCommentForPullRequest, which posts a comment to a pull request at the specified line, file, or request.

UpdatePullRequestDescription, which updates the description of a pull request.

UpdatePullRequestStatus, which updates the status of a pull request.

UpdatePullRequestTitle, which updates the title of a pull request.

Information about comments in a repository, by calling the following:

DeleteCommentContent, which deletes the content of a comment on a commit in a repository.

GetComment, which returns information about a comment on a commit.

GetCommentsForComparedCommit, which returns information about comments on the comparison between two commit specifiers in a repository.

PostCommentForComparedCommit, which creates a comment on the comparison between two commit specifiers in a repository.

PostCommentReply, which creates a reply to a comment.

UpdateComment, which updates the content of a comment on a commit in a repository.

Triggers, by calling the following:

GetRepositoryTriggers, which returns information about triggers configured for a repository.

PutRepositoryTriggers, which replaces all triggers for a repository and can be used to create or delete triggers.

TestRepositoryTriggers, which tests the functionality of a repository trigger by sending data to the trigger target.

For information about how to use AWS CodeCommit, see the AWS CodeCommit User Guide.

If you're using the service, you're probably looking for CodeCommitClient and CodeCommit.

Structs

BatchGetRepositoriesInput
Represents the input of a batch get repositories operation.

BatchGetRepositoriesOutput
Represents the output of a batch get repositories operation.

BlobMetadata
Returns information about a specific Git blob object.

BranchInfo
Returns information about a branch.

CodeCommitClient
A client for the CodeCommit API.

Comment
Returns information about a specific comment.

CommentsForComparedCommit
Returns information about comments on the comparison between two commits.

CommentsForPullRequest
Returns information about comments on a pull request.

Commit
Returns information about a specific commit.

CreateBranchInput
Represents the input of a create branch operation.

CreatePullRequestInput

CreatePullRequestOutput

CreateRepositoryInput
Represents the input of a create repository operation.

CreateRepositoryOutput
Represents the output of a create repository operation.

DeleteBranchInput
Represents the input of a delete branch operation.

DeleteBranchOutput
Represents the output of a delete branch operation.

DeleteCommentContentInput

DeleteCommentContentOutput

DeleteRepositoryInput
Represents the input of a delete repository operation.

DeleteRepositoryOutput
Represents the output of a delete repository operation.

DescribePullRequestEventsInput

DescribePullRequestEventsOutput

Difference
Returns information about a set of differences for a commit specifier.

GetBlobInput
Represents the input of a get blob operation.

GetBlobOutput
Represents the output of a get blob operation.

GetBranchInput
Represents the input of a get branch operation.

GetBranchOutput
Represents the output of a get branch operation.

GetCommentInput

GetCommentOutput

GetCommentsForComparedCommitInput

GetCommentsForComparedCommitOutput

GetCommentsForPullRequestInput

GetCommentsForPullRequestOutput

GetCommitInput
Represents the input of a get commit operation.

GetCommitOutput
Represents the output of a get commit operation.

GetDifferencesInput

GetDifferencesOutput

GetMergeConflictsInput

GetMergeConflictsOutput

GetPullRequestInput

GetPullRequestOutput

GetRepositoryInput
Represents the input of a get repository operation.

GetRepositoryOutput
Represents the output of a get repository operation.

GetRepositoryTriggersInput
Represents the input of a get repository triggers operation.

GetRepositoryTriggersOutput
Represents the output of a get repository triggers operation.

ListBranchesInput
Represents the input of a list branches operation.

ListBranchesOutput
Represents the output of a list branches operation.

ListPullRequestsInput

ListPullRequestsOutput

ListRepositoriesInput
Represents the input of a list repositories operation.

ListRepositoriesOutput
Represents the output of a list repositories operation.

Location
Returns information about the location of a change or comment in the comparison between two commits or a pull request.

MergeMetadata
Returns information about a merge or potential merge between a source reference and a destination reference in a pull request.

MergePullRequestByFastForwardInput

MergePullRequestByFastForwardOutput

PostCommentForComparedCommitInput

PostCommentForComparedCommitOutput

PostCommentForPullRequestInput

PostCommentForPullRequestOutput

PostCommentReplyInput

PostCommentReplyOutput

PullRequest
Returns information about a pull request.

PullRequestEvent
Returns information about a pull request event.

PullRequestMergedStateChangedEventMetadata
Returns information about the change in the merge state for a pull request event.

PullRequestSourceReferenceUpdatedEventMetadata
Information about an update to the source branch of a pull request.

PullRequestStatusChangedEventMetadata
Information about a change to the status of a pull request.

PullRequestTarget
Returns information about a pull request target.

PutFileInput

PutFileOutput

PutRepositoryTriggersInput
Represents the input ofa put repository triggers operation.

PutRepositoryTriggersOutput
Represents the output of a put repository triggers operation.

RepositoryMetadata
Information about a repository.

RepositoryNameIdPair
Information about a repository name and ID.

RepositoryTrigger
Information about a trigger for a repository.

RepositoryTriggerExecutionFailure
A trigger failed to run.

Target
Returns information about a target for a pull request.

TestRepositoryTriggersInput
Represents the input of a test repository triggers operation.

TestRepositoryTriggersOutput
Represents the output of a test repository triggers operation.

UpdateCommentInput

UpdateCommentOutput

UpdateDefaultBranchInput
Represents the input of an update default branch operation.

UpdatePullRequestDescriptionInput

UpdatePullRequestDescriptionOutput

UpdatePullRequestStatusInput

UpdatePullRequestStatusOutput

UpdatePullRequestTitleInput

UpdatePullRequestTitleOutput

UpdateRepositoryDescriptionInput
Represents the input of an update repository description operation.

UpdateRepositoryNameInput
Represents the input of an update repository description operation.

UserInfo
Information about the user who made a specified commit.

Enums

BatchGetRepositoriesError
Errors returned by BatchGetRepositories

CreateBranchError
Errors returned by CreateBranch

CreatePullRequestError
Errors returned by CreatePullRequest

CreateRepositoryError
Errors returned by CreateRepository

DeleteBranchError
Errors returned by DeleteBranch

DeleteCommentContentError
Errors returned by DeleteCommentContent

DeleteRepositoryError
Errors returned by DeleteRepository

DescribePullRequestEventsError
Errors returned by DescribePullRequestEvents

GetBlobError
Errors returned by GetBlob

GetBranchError
Errors returned by GetBranch

GetCommentError
Errors returned by GetComment

GetCommentsForComparedCommitError
Errors returned by GetCommentsForComparedCommit

GetCommentsForPullRequestError
Errors returned by GetCommentsForPullRequest

GetCommitError
Errors returned by GetCommit

GetDifferencesError
Errors returned by GetDifferences

GetMergeConflictsError
Errors returned by GetMergeConflicts

GetPullRequestError
Errors returned by GetPullRequest

GetRepositoryError
Errors returned by GetRepository

GetRepositoryTriggersError
Errors returned by GetRepositoryTriggers

ListBranchesError
Errors returned by ListBranches

ListPullRequestsError
Errors returned by ListPullRequests

ListRepositoriesError
Errors returned by ListRepositories

MergePullRequestByFastForwardError
Errors returned by MergePullRequestByFastForward

PostCommentForComparedCommitError
Errors returned by PostCommentForComparedCommit

PostCommentForPullRequestError
Errors returned by PostCommentForPullRequest

PostCommentReplyError
Errors returned by PostCommentReply

PutFileError
Errors returned by PutFile

PutRepositoryTriggersError
Errors returned by PutRepositoryTriggers

TestRepositoryTriggersError
Errors returned by TestRepositoryTriggers

UpdateCommentError
Errors returned by UpdateComment

UpdateDefaultBranchError
Errors returned by UpdateDefaultBranch

UpdatePullRequestDescriptionError
Errors returned by UpdatePullRequestDescription

UpdatePullRequestStatusError
Errors returned by UpdatePullRequestStatus

UpdatePullRequestTitleError
Errors returned by UpdatePullRequestTitle

UpdateRepositoryDescriptionError
Errors returned by UpdateRepositoryDescription

UpdateRepositoryNameError
Errors returned by UpdateRepositoryName

Traits

CodeCommit
Trait representing the capabilities of the CodeCommit API. CodeCommit clients implement this trait.