Struct rusoto_codecommit::CodeCommitClient

pub struct CodeCommitClient { /* fields omitted */ }

A client for the CodeCommit API.

Methods

impl CodeCommitClient

pub fn new(region: Region) -> CodeCommitClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CodeCommitClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,