Trait rusoto_codecommit::CodeCommit

pub trait CodeCommit {
    fn batch_get_repositories(
        &self, 
        input: BatchGetRepositoriesInput
    ) -> RusotoFuture<BatchGetRepositoriesOutput, BatchGetRepositoriesError>;

    fn create_branch(
        &self, 
        input: CreateBranchInput
    ) -> RusotoFuture<(), CreateBranchError>;

    fn create_pull_request(
        &self, 
        input: CreatePullRequestInput
    ) -> RusotoFuture<CreatePullRequestOutput, CreatePullRequestError>;

    fn create_repository(
        &self, 
        input: CreateRepositoryInput
    ) -> RusotoFuture<CreateRepositoryOutput, CreateRepositoryError>;

    fn delete_branch(
        &self, 
        input: DeleteBranchInput
    ) -> RusotoFuture<DeleteBranchOutput, DeleteBranchError>;

    fn delete_comment_content(
        &self, 
        input: DeleteCommentContentInput
    ) -> RusotoFuture<DeleteCommentContentOutput, DeleteCommentContentError>;

    fn delete_repository(
        &self, 
        input: DeleteRepositoryInput
    ) -> RusotoFuture<DeleteRepositoryOutput, DeleteRepositoryError>;

    fn describe_pull_request_events(
        &self, 
        input: DescribePullRequestEventsInput
    ) -> RusotoFuture<DescribePullRequestEventsOutput, DescribePullRequestEventsError>;

    fn get_blob(
        &self, 
        input: GetBlobInput
    ) -> RusotoFuture<GetBlobOutput, GetBlobError>;

    fn get_branch(
        &self, 
        input: GetBranchInput
    ) -> RusotoFuture<GetBranchOutput, GetBranchError>;

    fn get_comment(
        &self, 
        input: GetCommentInput
    ) -> RusotoFuture<GetCommentOutput, GetCommentError>;

    fn get_comments_for_compared_commit(
        &self, 
        input: GetCommentsForComparedCommitInput
    ) -> RusotoFuture<GetCommentsForComparedCommitOutput, GetCommentsForComparedCommitError>;

    fn get_comments_for_pull_request(
        &self, 
        input: GetCommentsForPullRequestInput
    ) -> RusotoFuture<GetCommentsForPullRequestOutput, GetCommentsForPullRequestError>;

    fn get_commit(
        &self, 
        input: GetCommitInput
    ) -> RusotoFuture<GetCommitOutput, GetCommitError>;

    fn get_differences(
        &self, 
        input: GetDifferencesInput
    ) -> RusotoFuture<GetDifferencesOutput, GetDifferencesError>;

    fn get_merge_conflicts(
        &self, 
        input: GetMergeConflictsInput
    ) -> RusotoFuture<GetMergeConflictsOutput, GetMergeConflictsError>;

    fn get_pull_request(
        &self, 
        input: GetPullRequestInput
    ) -> RusotoFuture<GetPullRequestOutput, GetPullRequestError>;

    fn get_repository(
        &self, 
        input: GetRepositoryInput
    ) -> RusotoFuture<GetRepositoryOutput, GetRepositoryError>;

    fn get_repository_triggers(
        &self, 
        input: GetRepositoryTriggersInput
    ) -> RusotoFuture<GetRepositoryTriggersOutput, GetRepositoryTriggersError>;

    fn list_branches(
        &self, 
        input: ListBranchesInput
    ) -> RusotoFuture<ListBranchesOutput, ListBranchesError>;

    fn list_pull_requests(
        &self, 
        input: ListPullRequestsInput
    ) -> RusotoFuture<ListPullRequestsOutput, ListPullRequestsError>;

    fn list_repositories(
        &self, 
        input: ListRepositoriesInput
    ) -> RusotoFuture<ListRepositoriesOutput, ListRepositoriesError>;

    fn merge_pull_request_by_fast_forward(
        &self, 
        input: MergePullRequestByFastForwardInput
    ) -> RusotoFuture<MergePullRequestByFastForwardOutput, MergePullRequestByFastForwardError>;

    fn post_comment_for_compared_commit(
        &self, 
        input: PostCommentForComparedCommitInput
    ) -> RusotoFuture<PostCommentForComparedCommitOutput, PostCommentForComparedCommitError>;

    fn post_comment_for_pull_request(
        &self, 
        input: PostCommentForPullRequestInput
    ) -> RusotoFuture<PostCommentForPullRequestOutput, PostCommentForPullRequestError>;

    fn post_comment_reply(
        &self, 
        input: PostCommentReplyInput
    ) -> RusotoFuture<PostCommentReplyOutput, PostCommentReplyError>;

    fn put_file(
        &self, 
        input: PutFileInput
    ) -> RusotoFuture<PutFileOutput, PutFileError>;

    fn put_repository_triggers(
        &self, 
        input: PutRepositoryTriggersInput
    ) -> RusotoFuture<PutRepositoryTriggersOutput, PutRepositoryTriggersError>;

    fn test_repository_triggers(
        &self, 
        input: TestRepositoryTriggersInput
    ) -> RusotoFuture<TestRepositoryTriggersOutput, TestRepositoryTriggersError>;

    fn update_comment(
        &self, 
        input: UpdateCommentInput
    ) -> RusotoFuture<UpdateCommentOutput, UpdateCommentError>;

    fn update_default_branch(
        &self, 
        input: UpdateDefaultBranchInput
    ) -> RusotoFuture<(), UpdateDefaultBranchError>;

    fn update_pull_request_description(
        &self, 
        input: UpdatePullRequestDescriptionInput
    ) -> RusotoFuture<UpdatePullRequestDescriptionOutput, UpdatePullRequestDescriptionError>;

    fn update_pull_request_status(
        &self, 
        input: UpdatePullRequestStatusInput
    ) -> RusotoFuture<UpdatePullRequestStatusOutput, UpdatePullRequestStatusError>;

    fn update_pull_request_title(
        &self, 
        input: UpdatePullRequestTitleInput
    ) -> RusotoFuture<UpdatePullRequestTitleOutput, UpdatePullRequestTitleError>;

    fn update_repository_description(
        &self, 
        input: UpdateRepositoryDescriptionInput
    ) -> RusotoFuture<(), UpdateRepositoryDescriptionError>;

    fn update_repository_name(
        &self, 
        input: UpdateRepositoryNameInput
    ) -> RusotoFuture<(), UpdateRepositoryNameError>;
}

Trait representing the capabilities of the CodeCommit API. CodeCommit clients implement this trait.

Required Methods

fn batch_get_repositories(
&self,
input: BatchGetRepositoriesInput
) -> RusotoFuture<BatchGetRepositoriesOutput, BatchGetRepositoriesError>

Returns information about one or more repositories.

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

fn create_branch(
&self,
input: CreateBranchInput
) -> RusotoFuture<(), CreateBranchError>

Creates a new branch in a repository and points the branch to a commit.

Calling the create branch operation does not set a repository's default branch. To do this, call the update default branch operation.

fn create_pull_request(
&self,
input: CreatePullRequestInput
) -> RusotoFuture<CreatePullRequestOutput, CreatePullRequestError>

Creates a pull request in the specified repository.

fn create_repository(
&self,
input: CreateRepositoryInput
) -> RusotoFuture<CreateRepositoryOutput, CreateRepositoryError>

Creates a new, empty repository.

fn delete_branch(
&self,
input: DeleteBranchInput
) -> RusotoFuture<DeleteBranchOutput, DeleteBranchError>

Deletes a branch from a repository, unless that branch is the default branch for the repository.

fn delete_comment_content(
&self,
input: DeleteCommentContentInput
) -> RusotoFuture<DeleteCommentContentOutput, DeleteCommentContentError>

Deletes the content of a comment made on a change, file, or commit in a repository.

fn delete_repository(
&self,
input: DeleteRepositoryInput
) -> RusotoFuture<DeleteRepositoryOutput, DeleteRepositoryError>

Deletes a repository. If a specified repository was already deleted, a null repository ID will be returned.

Deleting a repository also deletes all associated objects and metadata. After a repository is deleted, all future push calls to the deleted repository will fail.

fn describe_pull_request_events(
&self,
input: DescribePullRequestEventsInput
) -> RusotoFuture<DescribePullRequestEventsOutput, DescribePullRequestEventsError>

Returns information about one or more pull request events.

fn get_blob(
&self,
input: GetBlobInput
) -> RusotoFuture<GetBlobOutput, GetBlobError>

Returns the base-64 encoded content of an individual blob within a repository.

fn get_branch(
&self,
input: GetBranchInput
) -> RusotoFuture<GetBranchOutput, GetBranchError>

Returns information about a repository branch, including its name and the last commit ID.

fn get_comment(
&self,
input: GetCommentInput
) -> RusotoFuture<GetCommentOutput, GetCommentError>

Returns the content of a comment made on a change, file, or commit in a repository.

fn get_comments_for_compared_commit(
&self,
input: GetCommentsForComparedCommitInput
) -> RusotoFuture<GetCommentsForComparedCommitOutput, GetCommentsForComparedCommitError>

Returns information about comments made on the comparison between two commits.

fn get_comments_for_pull_request(
&self,
input: GetCommentsForPullRequestInput
) -> RusotoFuture<GetCommentsForPullRequestOutput, GetCommentsForPullRequestError>

Returns comments made on a pull request.

fn get_commit(
&self,
input: GetCommitInput
) -> RusotoFuture<GetCommitOutput, GetCommitError>

Returns information about a commit, including commit message and committer information.

fn get_differences(
&self,
input: GetDifferencesInput
) -> RusotoFuture<GetDifferencesOutput, GetDifferencesError>

Returns information about the differences in a valid commit specifier (such as a branch, tag, HEAD, commit ID or other fully qualified reference). Results can be limited to a specified path.

fn get_merge_conflicts(
&self,
input: GetMergeConflictsInput
) -> RusotoFuture<GetMergeConflictsOutput, GetMergeConflictsError>

Returns information about merge conflicts between the before and after commit IDs for a pull request in a repository.

fn get_pull_request(
&self,
input: GetPullRequestInput
) -> RusotoFuture<GetPullRequestOutput, GetPullRequestError>

Gets information about a pull request in a specified repository.

fn get_repository(
&self,
input: GetRepositoryInput
) -> RusotoFuture<GetRepositoryOutput, GetRepositoryError>

Returns information about a repository.

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

fn get_repository_triggers(
&self,
input: GetRepositoryTriggersInput
) -> RusotoFuture<GetRepositoryTriggersOutput, GetRepositoryTriggersError>

Gets information about triggers configured for a repository.

fn list_branches(
&self,
input: ListBranchesInput
) -> RusotoFuture<ListBranchesOutput, ListBranchesError>

Gets information about one or more branches in a repository.

fn list_pull_requests(
&self,
input: ListPullRequestsInput
) -> RusotoFuture<ListPullRequestsOutput, ListPullRequestsError>

Returns a list of pull requests for a specified repository. The return list can be refined by pull request status or pull request author ARN.

fn list_repositories(
&self,
input: ListRepositoriesInput
) -> RusotoFuture<ListRepositoriesOutput, ListRepositoriesError>

Gets information about one or more repositories.

fn merge_pull_request_by_fast_forward(
&self,
input: MergePullRequestByFastForwardInput
) -> RusotoFuture<MergePullRequestByFastForwardOutput, MergePullRequestByFastForwardError>

Closes a pull request and attempts to merge the source commit of a pull request into the specified destination branch for that pull request at the specified commit using the fast-forward merge option.

fn post_comment_for_compared_commit(
&self,
input: PostCommentForComparedCommitInput
) -> RusotoFuture<PostCommentForComparedCommitOutput, PostCommentForComparedCommitError>

Posts a comment on the comparison between two commits.

fn post_comment_for_pull_request(
&self,
input: PostCommentForPullRequestInput
) -> RusotoFuture<PostCommentForPullRequestOutput, PostCommentForPullRequestError>

Posts a comment on a pull request.

fn post_comment_reply(
&self,
input: PostCommentReplyInput
) -> RusotoFuture<PostCommentReplyOutput, PostCommentReplyError>

Posts a comment in reply to an existing comment on a comparison between commits or a pull request.

fn put_file(
&self,
input: PutFileInput
) -> RusotoFuture<PutFileOutput, PutFileError>

Adds or updates a file in an AWS CodeCommit repository.

fn put_repository_triggers(
&self,
input: PutRepositoryTriggersInput
) -> RusotoFuture<PutRepositoryTriggersOutput, PutRepositoryTriggersError>

Replaces all triggers for a repository. This can be used to create or delete triggers.

fn test_repository_triggers(
&self,
input: TestRepositoryTriggersInput
) -> RusotoFuture<TestRepositoryTriggersOutput, TestRepositoryTriggersError>

Tests the functionality of repository triggers by sending information to the trigger target. If real data is available in the repository, the test will send data from the last commit. If no data is available, sample data will be generated.

fn update_comment(
&self,
input: UpdateCommentInput
) -> RusotoFuture<UpdateCommentOutput, UpdateCommentError>

Replaces the contents of a comment.

fn update_default_branch(
&self,
input: UpdateDefaultBranchInput
) -> RusotoFuture<(), UpdateDefaultBranchError>

Sets or changes the default branch name for the specified repository.

If you use this operation to change the default branch name to the current default branch name, a success message is returned even though the default branch did not change.

fn update_pull_request_description(
&self,
input: UpdatePullRequestDescriptionInput
) -> RusotoFuture<UpdatePullRequestDescriptionOutput, UpdatePullRequestDescriptionError>

Replaces the contents of the description of a pull request.

fn update_pull_request_status(
&self,
input: UpdatePullRequestStatusInput
) -> RusotoFuture<UpdatePullRequestStatusOutput, UpdatePullRequestStatusError>

Updates the status of a pull request.

fn update_pull_request_title(
&self,
input: UpdatePullRequestTitleInput
) -> RusotoFuture<UpdatePullRequestTitleOutput, UpdatePullRequestTitleError>

Replaces the title of a pull request.

fn update_repository_description(
&self,
input: UpdateRepositoryDescriptionInput
) -> RusotoFuture<(), UpdateRepositoryDescriptionError>

Sets or changes the comment or description for a repository.

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

fn update_repository_name(
&self,
input: UpdateRepositoryNameInput
) -> RusotoFuture<(), UpdateRepositoryNameError>

Renames a repository. The repository name must be unique across the calling AWS account. In addition, repository names are limited to 100 alphanumeric, dash, and underscore characters, and cannot include certain characters. The suffix ".git" is prohibited. For a full description of the limits on repository names, see Limits in the AWS CodeCommit User Guide.